Settle on a rule for attributes
It's going to be this: only reindent if there was a newline.

It's too hard to get right, the rules would depend on the attribute.
e.g.

- `class` would probably want to split on space
- `srcset`, `sizes` would want to split on comma
- `style`, would want to split on semi-colon
- `alt`, would want paragraph style splitting
- `href`, `id`, `action`, `method`, etc. probably don't want line breaks at all

Then there's `data-X` for folks using lazysizes and the like...

So for that reason, I think the only sane thing to do is to break on a
new line and reindent if there was a new line in the source, and
otherwise don't format it at all.

This should be alright.