Extract helpers out of LiquidHTML grammar
This is phase one, we'll be able to move Liquid and LiquidStatement out
in two different grammars next.

Once we have those, we'll be able to grab the `liquid` tag markup and
make it go through the _other_ grammar but with similar mappings.

Something like:

- Helpers
- Liquid <: Helpers
- LiquidStatement <: Liquid
- LiquidHTML <: Liquid

And in LiquidStatement, we can do something big brain like:

```ohm
LiquidStatement <: Liquid {
  statements = listOf<statement, statementSep>
  statementSep = newline whitespace*
  statement =
    | liquidEchoStatement
    | liquidAssignStatement
    | /* ... */
  space = " " | "\t"
  whitespace = (newline | space)
}
```

And then in the mapping for liquidTagLiquidMarkup something like
```js
liquidTagLiquidMarkup(markup) {
  const res = liquidStatementGrammar.parse(markup, 'statements');
  /* ... */
}
```