Make the CST responsible for setting `.source`, clean up args in stage-2
The whole `.source` thing was kind of awkward and caused all the methods
in `stage-2-ast` to have a `source` argument.

This commit moves the `.source` to the CST generation phase. This way,
all CST nodes have a `.source` so we don't have to pass that argument
around so much.

The only reason it was there was so we could properly throw LiquidHtmlASTParsingError
which depends on the source.

Thank god for TypeScript for this.